+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Anticipated release: December 21, 2020
 
 #### 🚀 New features
 
-- Update endpoint for affiliations to filter by status ([#2682])
+- Update endpoint for affiliations to filter by status ([#2692])
 
 #### 🐛 Bugs fixed
 
@@ -20,7 +20,8 @@ Anticipated release: December 21, 2020
 
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
+[#2692]: https://github.com/CMSgov/eAPD/issues/2692
 [#2348]: https://github.com/CMSgov/eAPD/issues/2348
 [#2682]: https://github.com/CMSgov/eAPD/issues/2682
-[GHSA-vrv8-v4w8-f95h]: https://github.com/advisories/GHSA-vrv8-v4w8-f95h
-[GHSA-w7rc-rwvf-8q5r]: https://github.com/advisories/GHSA-w7rc-rwvf-8q5r
+[ghsa-vrv8-v4w8-f95h]: https://github.com/advisories/GHSA-vrv8-v4w8-f95h
+[ghsa-w7rc-rwvf-8q5r]: https://github.com/advisories/GHSA-w7rc-rwvf-8q5r